Até resultfragment, sem adicionar foto ao banco.
Com banco de dados, ainda sem corrigir erros do método toList e sem
navegabilidade.

+++ FaunaRS-1/app/src/main/java/model/Galeria.java
@@ -8,7 +8,7 @@ import java.util.Arrays;
  */
 public class Galeria implements Serializable{
     public static final long serialVersionUID = 1L;
-    public Long gal_id, esp_id;
+    public Integer gal_id, esp_id;
     public byte[] imagem;
 
     public Galeria() {
